Only download database assets in development environment if they do not exist
Any image built in CI should not have these assets present on filesystem
so should always download fresh up-to-date version

+++ bin/load.sh
@@ -14,8 +14,12 @@ for F in *.csv; do
   rm "$F"
 done
 
-echo "downloading from $URL"
-curl -O $URL
+if [ ! -f "$(basename "$URL")" ]; then
+  echo "downloading from $URL"
+  curl -O $URL
+else
+  echo "using $(basename "$URL") from local filesystem"
+fi
 
 echo "exporting"
